% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_assessment.R
\name{dataset_assessment}
\alias{dataset_assessment}
\title{Return formatted assessment dataset}
\usage{
dataset_assessment(
  module = c("All", "AAA", "BBB", "CCC", "DDD", "EEE", "FFF", "GGG"),
  presentation = c("2013B", "2014B", "2013J", "2014J", "All", "Summer", "Winter"),
  repeat_students = c("remove", "keep")
)
}
\arguments{
\item{module}{Name of the module to be included, either \code{"AAA"}, \code{"BBB"}, \code{"CCC"}, \code{"DDD"}, \code{"EEE"}, \code{"FFF"} or \code{"GGG"}.}

\item{presentation}{Name of the semester of the module to be included, either \code{"2013B"},
\code{"2014B"}, \code{"2013J"}, \code{"2014J"}, \code{"All"}, \code{"Summer"} or \code{"Winter"}.
\code{"B"} indicates a February start time whereas \code{"J"} indicates an October start time. \code{"All"} indicates that all presentations of the module will be included in the returned data.}

\item{repeat_students}{Indicator of whether students who had previous attempts at the module should be removed, either \code{"remove"} or \code{"keep"}.}
}
\value{
Returns the inputs specified - module, presentation and whether repeat students are to be included.
Four tibbles are also returned: 1) assessment_data, 2) assessment_performance,
3) assessment_reactivity and 4) assessments.
}
\description{
Load, combines and formats the assessment and student assessment datasets from the OULAD for data analysis.
}
\section{assessment_data tibble}{

A tibble based on the combined oulad files of studentAssessment.csv and assessments.csv,
and the inputs. The tibble consists of (Kuzilek et al., 2017):

\itemize{
\item{id_assessment - the assessment identification number.}
\item{code_module - the module identification code.}
\item{code_presentation - the presentation identification code.}
\item{assessment_type - the type of assessment.}
\item{date - due date of the assessment in the module term.}
\item{weight - the weight of the assessment.}
\item{id_student - the unique student identification number.}
\item{date_submitted - the day of the assessment submission by student.}
\item{is_banked - logical indicator whether assessment result has been transferred from a previous presentation.}
\item{score - the student’s score for the specific assessment (range 0-100).}
\item{reactivity - calculated using the date variable minus date_submitted variable.}
}
}

\section{assessments}{

A tibble of six columns which details for each assessment item for a module presentation,
the type of assessment it is, the due date of the assessment,
and the weight of the assessment. The assessment types are Tutor Marked Assessment (TMA), Computer Marked Assessment (CMA)
and Final Exam (Exam). Exams are generally treated separately to other assessments
and have weight equal to 100\%. The sum of all other assessments is also 100\% (Kuzilek et al., 2017).
}

\section{assessment_performance tibble}{

A tibble where each row represents a unique student and their scores in the range of 0-100
for different assessment items. The final column states each student's average assessment score based on their assessment
scores and the assessments weights. For this calculation NA is treated as 0.
Details of the assessment items can be seen in the assessments tibble.
}

\section{assessment_reactivity}{

A tibble where each row represents a unique student and their reactivity for
different assessment items. Treuillier and Boyer (2021) define reactivity for the oulad as the delay between
the date the assessment is returned and the deadline (in days). Negative numbers indicate overdue assessments.
Details of the assessment items can be seen in the assessments tibble.
}

\examples{
dataset_assessment(module = "BBB", presentation = "2013J", repeat_students = "remove")
}
\references{
Treuillier, C., & Boyer, A. (2021). Identification of class-representative learner personas.
In LA4SLE 2021 - Learning Analytics for Smart Learning Environments (pp. 38-45). Bolzano, Italy.
⟨hal-03549915

Kuzilek, J., Hlosta, M., & Zdrahal, Z. (2017). Open university learning analytics dataset. Scientific Data
volume 4, (pp. 1–8). https://doi.org/10.1038/sdata.2017.171.
}
\seealso{
\code{\link{combined_dataset}}
}
