% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_VLE_activity.R
\name{dataset_VLE_activity}
\alias{dataset_VLE_activity}
\title{Return formatted VLE dataset based on activity types}
\usage{
dataset_VLE_activity(
  module = c("AAA", "BBB", "CCC", "DDD", "EEE", "FFF", "GGG"),
  presentation = c("2013B", "2014B", "2013J", "2014J", "All"),
  repeat_students = c("remove", "keep"),
  week_begin = -4,
  week_end = 39,
  example_data = FALSE
)
}
\arguments{
\item{module}{Name of the module to be included, either \code{"AAA"}, \code{"BBB"}, \code{"CCC"}, \code{"DDD"}, \code{"EEE"}, \code{"FFF"} or \code{"GGG"}.}

\item{presentation}{Name of the semester of the module to be included, either \code{"2013B"},
\code{"2014B"}, \code{"2013J"}, \code{"2014J"}, or \code{"All"}.
\code{"B"} indicates a February start time whereas \code{"J"} indicates an October start time. \code{"All"} indicates that all presentations of the module will be included in the returned data.}

\item{repeat_students}{Indicator of whether students who had previous attempts at the module should be removed, either \code{"remove"} or \code{"keep"}.
If presentation is set to \code{"All"}, this is automatically set to \code{"remove"}.}

\item{week_begin}{First week of VLE data to be included in formatted data. Depending on the presentation, students
started to view activities four weeks prior to the initial module start date. Weeks prior to the initial module start
are indicated by a negative integer.}

\item{week_end}{Last week of VLE data to be included in the formatted data.}

\item{example_data}{TRUE/FALSE indicator for whether to run a subset of the data as an example}
}
\value{
Returns the inputs specified - module, presentation, whether repeat students are to be included, the first week of VLE data to be included
and the last week of VLE data to be included.
Two tibbles are also returned: 1) filtered_data, and 2) resource_data.
}
\description{
Load and formats the student Virtual Learning Environment (VLE) dataset from the OULAD for data analysis.
}
\section{filtered_data tibble}{

A tibble based on the combined oulad files of studentVLE.csv and vle.csv,
and the inputs. The tibble consists of (Kuzilek et al., 2017):

\itemize{
\item{code_module - the module identification code.}
\item{code_presentation - the presentation identification code.}
\item{id_site - the VLE material identification number.}
\item{activity_type - the role associated with the module material.}
\item{week_from - the week from which the material is planned to be used.}
\item{week_to - the week until which the material is planned to be used.}
\item{id_student - the unique student identification number.}
\item{date - the day of student’s interaction with the material.}
\item{sum_click - the number of times the student interacted with the material.}
}
}

\section{resource_data tibble}{

A tibble where each row represents a unique student and their number of interactions with
different VLE activity types for the period (weeks) and module presentation inputted.
}

\examples{
dataset_VLE_activity(module = "AAA", presentation = "2013J", repeat_students = "remove",
                            week_begin = 1, week_end = 39, example_data = TRUE)
}
\references{
Kuzilek, J., Hlosta, M., & Zdrahal, Z. (2017). Open university learning analytics dataset. Scientific Data
volume 4 , (pp. 1–8). https://doi.org/10.1038/sdata.2017.171.
}
\seealso{
\code{\link{convert_VLE}}, \code{\link{dataset_VLE_time}}, \code{\link{VLE_learning_classification}} and \code{\link{combined_dataset}}
}
