% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_VLE_time.R
\name{dataset_VLE_time}
\alias{dataset_VLE_time}
\title{Return formatted VLE dataset for daily and weekly counts}
\usage{
dataset_VLE_time(
  module = c("AAA", "BBB", "CCC", "DDD", "EEE", "FFF", "GGG"),
  presentation = c("2013B", "2014B", "2013J", "2014J", "All"),
  repeat_students = c("remove", "keep"),
  example_data = FALSE
)
}
\arguments{
\item{module}{Name of the module to be included, either \code{"AAA"}, \code{"BBB"}, \code{"CCC"}, \code{"DDD"}, \code{"EEE"}, \code{"FFF"} or \code{"GGG"}.}

\item{presentation}{Name of the semester of the module to be included, either \code{"2013B"},
\code{"2014B"}, \code{"2013J"}, \code{"2014J"}, or \code{"All"}.
\code{"B"} indicates a February start time whereas \code{"J"} indicates an October start time. \code{"All"} indicates that all presentations of the module will be included in the returned data.}

\item{repeat_students}{Indicator of whether students who had previous attempts at the module should be removed, either \code{"remove"} or \code{"keep"}.
If presentation is set to \code{"All"}, this is automatically set to \code{"remove"}.}

\item{example_data}{TRUE/FALSE indicator for whether to run a subset of the data as an example.}
}
\value{
Returns the inputs specified - module, presentation, and whether repeat students are to be included.
Three tibbles are also returned: 1) filtered_data, 2) daily_data, and 3) weekly_data.
}
\description{
Load and formats the student Virtual Learning Environment (VLE) dataset from the OULAD for data analysis.
}
\section{filtered_data tibble}{

A tibble based on the students' VLE interactions and the inputs. The tibble consists of (Kuzilek et al., 2017):

\itemize{
\item{id_student - the unique student identification number.}
\item{date - the day of the student’s interaction with the material.}
\item{sum_click - the number of times the student interacted with the material on a specific day.}
}
}

\section{daily_data}{

A tibble where each row represents a unique student and their number of interactions with
the VLE for different days of the module. The module starts on day 0. Days prior to the initial start day, are indicated by negative integers.
}

\section{weekly_data}{

A tibble where each row represents a unique student and their number of interactions with
the VLE for different weeks of the module. Weeks prior to the initial start week (week 1) are indicated by the term 'pre' in the week name.
}

\examples{
VLE_data = dataset_VLE_time(module = "AAA", presentation = "2013J",
repeat_students = "remove", example_data = TRUE)
}
\references{
Kuzilek, J., Hlosta, M., & Zdrahal, Z. (2017). Open university learning analytics dataset. Scientific Data
volume 4 , (pp. 1–8). https://doi.org/10.1038/sdata.2017.171.
}
\seealso{
\code{\link{convert_VLE}}, \code{\link{dataset_VLE_activity}}, and \code{\link{combined_dataset}}
}
