% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_dataset.R
\name{combined_dataset}
\alias{combined_dataset}
\title{Return combined OULAD datasets}
\usage{
combined_dataset(
  module = c("AAA", "BBB", "CCC", "DDD", "EEE", "FFF", "GGG"),
  presentation = c("2013J", "2014J", "2013B", "2014B", "All"),
  repeat_students = c("remove", "keep"),
  demographics = c("include", "omit"),
  assessment = c("include", "omit"),
  registration = c("include", "omit"),
  VLE = c("daily", "weekly", "activity", "FSLM", "FSLSM", "OLS", "VARK", "omit"),
  VLE_clicks = c("total", "binary", "standardise", "logarithmic"),
  week_begin = 1,
  week_end = 39,
  example_data = FALSE
)
}
\arguments{
\item{module}{Name of the module to be included, either \code{"AAA"}, \code{"BBB"}, \code{"CCC"},
\code{"DDD"}, \code{"EEE"}, \code{"FFF"} or \code{"GGG"}.}

\item{presentation}{Name of the semester of the module to be included, either \code{"2013B"},
\code{"2014B"}, \code{"2013J"}, \code{"2014J"}, or \code{"All"}.
\code{"B"} indicates a February start time whereas \code{"J"} indicates an October start time. \code{"All"} indicates that all presentations of the module will be included in the returned data.}

\item{repeat_students}{Whether students who had previous attempts at the module should be removed, either \code{"remove"} or \code{"keep"}.
When presentation is set to \code{"All"}, this value is set to \code{"remove"}.}

\item{demographics}{Indicates whether demographic data is included in the outputted dataset.}

\item{assessment}{Indicates whether assessment data is included in the outputted dataset.}

\item{registration}{Indicates whether registration data is included in the outputted dataset.}

\item{VLE}{Indicates whether VLE data is included in the outputted dataset (\code{"omit"} if not included).
If included, the VLE data can returned based on \code{"weekly"} VLE interactions or \code{"daily"} VLE
interactions or interactions according to OU activities (\code{"activity"}).
The VLE data can also be returned as classified under a
specific learning model, either \code{"FSLM"}, \code{"FSLSM"}, \code{"OLS"}, or \code{"VARK"}.}

\item{VLE_clicks}{Indicates the format that the VLE data should be returned as, either \code{"total"} views,
\code{"binary"} views, \code{"standardise"} views using the scale function, or \code{"logarithmic"} values.}

\item{week_begin}{If VLE is based on OU activities, the first week of VLE data to be included in ouputted data. Depending on the presentation, students
started to view activities four weeks prior to the initial module start date. Weeks prior to the initial module start
are indicated by a negative integer.}

\item{week_end}{If VLE is based on OU activities, the last week of VLE data to be included in the outputted data.}

\item{example_data}{TRUE/FALSE indicator for whether to run a subset of the data as an example}
}
\value{
Returns the inputs specified for whether assessment, demographics, registration and VLE variables are to be included.
Also one tibble is returned, dataset_combined.
}
\description{
Combine multiple OULAD datasets into one tibble that is formatted for data analysis and
where each row represents a unique student
}
\section{dataset_combined}{

A tibble where each row represents a unique student. Depending on the inputs specified,
the tibble includes assessment, demographics, registration and VLE data for the student.
}

\examples{
combined_dataset(module = "AAA", presentation = "All",
repeat_students = "remove", demographics = "include",
assessment = "include", registration = "include",
VLE = "weekly", VLE_clicks = "total",
example_data = TRUE)

\dontrun{
combined_dataset(module = "BBB", presentation = "2013J",
repeat_students = "remove", demographics = "include",
assessment = "include", registration = "omit",
VLE = "activity", VLE_clicks = "binary",
week_begin = 1, week_end = 10, example_data = FALSE)}
}
\references{
Kuzilek, J., Hlosta, M., & Zdrahal, Z. (2017). Open university learning analytics dataset. Scientific Data
volume 4 , (pp. 1–8). https://doi.org/10.1038/sdata.2017.171.
}
\seealso{
For more information on different variable groupings see:
\itemize{
\item{ \code{\link{dataset_assessment}} for information on assessment performance data,}
\item{\code{\link{dataset_demographics}} for information on demographics data,}
\item{\code{\link{dataset_registration}} for information on registration data,}
\item{\code{\link{dataset_VLE_time}} for information on VLE daily or weekly data,}
\item{\code{\link{dataset_VLE_activity}} for information on VLE activity data,}
\item{\code{\link{VLE_learning_classification}} for information on VLE activities classified under a learning model, and}
\item{\code{\link{convert_VLE}} for information on transforming the type of VLE data.}
}
}
