% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_VLE.R
\name{convert_VLE}
\alias{convert_VLE}
\title{Converts the VLE data format}
\usage{
convert_VLE(data, conversion = c("binary", "standardise", "logarithmic"))
}
\arguments{
\item{data}{VLE data set to be converted. The VLE data set should consist of student_id
as the first column followed by total view counts per time period/activity in
each of the remaining columns.}

\item{conversion}{type of conversion to implement, either \code{"binary"}, \code{"standardise"} (using the scale function),
or \code{"logarithmic"}.}
}
\value{
Two tibbles are returned: 1) original_data, and 2) converted_data.
}
\description{
Converts the data format of the Virtual Learning Environment (VLE) data set from total view counts to binary, standardised or logarithmic view count data.
}
\section{original_data tibble}{

A tibble of the original data inputted into the function.
}

\section{converted_data tibble}{

A tibble where all columns, except the first column, have been changed from total view counts to either binary, standardised or logarithmic view count data.
}

\examples{
VLE_data = dataset_VLE_activity(example_data = TRUE)$resource_data
convert_VLE(VLE_data, conversion = "standardise")

\dontrun{
VLE_data = dataset_VLE_time(example_data = TRUE)$weekly_data
convert_VLE(VLE_data, conversion = "binary")

VLE_data = dataset_VLE_time(module = "BBB",
presentation = "2013J", repeat_students = "remove",
week_begin = 1, week_end = 13, example_data=FALSE)$weekly_data
convert_VLE(VLE_data, conversion = "binary")}
}
\seealso{
\code{\link{dataset_VLE_time}} or \code{\link{dataset_VLE_activity}} for obtaining the VLE data set needed for the input data.
}
